+++ tests/app.js
@@ -29,7 +29,7 @@ describe('generator-nikita:app', () => {
             jsFramework: 'jsb',
             nikitaCssMixins: [],
             nikitaCssExtends: [],
-            addons: [],
+            libraries: [],
         }));
 
     test('creates files', () => {
@@ -42,6 +42,12 @@ describe('generator-nikita:app', () => {
             'src/scss/styles.scss',
             'src/html/pages/index.twig',
             'static/img/appicons/favicon.ico',
+        ]);
+
+        assert.noFile([
+            'src/components/slider/Slider.jsb.js',
+            'src/components/select/Select.jsb.js',
+            'src/components/dialog/Dialog.jsb.js',
         ]);
     });
 });
@@ -60,7 +66,13 @@ describe('generator-nikita:web-app-jsb', () => {
             jsFramework: 'jsb',
             nikitaCssMixins: [],
             nikitaCssExtends: [],
-            addons: [],
+            libraries: [
+                'lodash',
+                'date-fns',
+                'choices',
+                'siema',
+                'a11y-dialog',
+            ],
         }));
 
     test('creates files', (done) => {
@@ -72,9 +84,17 @@ describe('generator-nikita:web-app-jsb', () => {
             'src/js/_main.js',
             'src/js/app.js',
             'src/components/sample/Sample.jsb.js',
+            'src/components/slider/Slider.jsb.js',
+            'src/components/select/Select.jsb.js',
+            'src/components/dialog/Dialog.jsb.js',
             'src/scss/styles.scss',
             'src/html/pages/index.twig',
             'static/img/appicons/favicon.ico',
+        ]);
+
+        assert.fileContent([
+            ['src/components/sample/Sample.jsb.js', /lodash-es/],
+            ['src/components/sample/Sample.jsb.js', /date-fns\/esm/],
         ]);
 
         if (process.env.TEMPLATE === 'web-app') {
@@ -101,7 +121,13 @@ describe('generator-nikita:web-app-react', () => {
             jsFramework: 'react',
             nikitaCssMixins: [],
             nikitaCssExtends: [],
-            addons: [],
+            libraries: [
+                'lodash',
+                'date-fns',
+                'siema',
+                'react-select',
+                'react-a11y-dialog',
+            ],
         }));
 
     test('creates files', (done) => {
@@ -115,9 +141,17 @@ describe('generator-nikita:web-app-react', () => {
             'src/js/Store.js',
             'src/components/sample/Sample.js',
             'src/components/counter/Counter.js',
+            'src/components/slider/Slider.js',
+            'src/components/select/Select.js',
+            'src/components/dialog/Dialog.js',
             'src/scss/styles.scss',
             'src/html/pages/index.twig',
             'static/img/appicons/favicon.ico',
+        ]);
+
+        assert.fileContent([
+            ['src/components/sample/Sample.js', /lodash-es/],
+            ['src/components/sample/Sample.js', /date-fns\/esm/],
         ]);
 
         if (process.env.TEMPLATE === 'web-app') {